docker/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary>=2.9.11
